# internal/app/api/core/assets/doc/v0_swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "WireGuard Portal API - UI Endpoints",
        "title": "WireGuard Portal SPA-UI API",
        "contact": {
            "name": "WireGuard Portal Developers",
            "url": "https://github.com/biezax/wg-portal"
        },
        "version": "0.0"
    },
    "basePath": "/api/v0",
    "paths": {
...
        "/auth/session": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Get information about the currently logged-in user.",
                "operationId": "auth_handleSessionInfoGet",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.SessionInfo"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
        "/auth/webauthn/credential/{id}": {
            "put": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Update a WebAuthn credential.",
                "operationId": "auth_handleWebAuthnCredentialsPut",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Base64 encoded Credential ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Credential name",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.WebAuthnCredentialRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.WebAuthnCredentialResponse"
                            }
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Delete a WebAuthn credential.",
                "operationId": "auth_handleWebAuthnCredentialsDelete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Base64 encoded Credential ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.WebAuthnCredentialResponse"
                            }
                        }
                    }
                }
            }
        },
...
        "/auth/webauthn/register/finish": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Finish the WebAuthn registration process.",
                "operationId": "auth_handleWebAuthnRegisterFinish",
                "parameters": [
                    {
                        "type": "string",
                        "default": "\"\"",
                        "description": "Credential name",
                        "name": "credential_name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.WebAuthnCredentialResponse"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Error": {
            "type": "object",
            "properties": {
                "Code": {
                    "type": "integer"
                },
                "Message": {
                    "type": "string"
                }
            }
        },
...
        "model.SessionInfo": {
            "type": "object",
            "properties": {
                "IsAdmin": {
                    "type": "boolean"
                },
                "LoggedIn": {
                    "type": "boolean"
                },
                "UserEmail": {
                    "type": "string"
                },
                "UserFirstname": {
                    "type": "string"
                },
                "UserIdentifier": {
                    "type": "string"
                },
                "UserLastname": {
                    "type": "string"
                }
            }
        },
...
        "model.WebAuthnCredentialRequest": {
            "type": "object",
            "properties": {
                "Name": {
                    "type": "string"
                }
            }
        },
        "model.WebAuthnCredentialResponse": {
            "type": "object",
            "properties": {
                "CreatedAt": {
                    "type": "string"
                },
                "ID": {
                    "type": "string"
                },
                "Name": {
                    "type": "string"
                }
            }
        }
    }
}